# ENH: optimize: vectorized minimization of univariate functions