… scipy/optimize/_zeros_py.py
… scipy/optimize/meson.build
… scipy/optimize/tests/meson.build
… scipy/optimize/tests/test_chandrupatla.py
… scipy/optimize/tests/test_zeros.py