@mdhaber 2023-07-05
> Am I missing something? 

This is a minimizer; the other is a root finder.